don't clear Objects position when Cartesian is set
When set<Cartesian> was called, the position was cleared if Cartesian
was defaulted.

This meant that this would have a default position rather than the one
specified:

    AudioBlockFormatObjects(CartesianPosition(...), Cartesian(true))

Because set<CartesianPosition> also calls set<Cartesian>, this would
have the same effect:

    AudioBlockFormatObjects(CartesianPosition(...))

To solve this, set<> for Cartesian, SphericalPosition and
CartesianPosition now all just manipulate cartesian_, sphericalPosition_
and cartesianPosition_ directly, rather than calling each other.

This isn't the smallest change to fix this, but it does make it
obviously correct -- there is no infinite recursion possible here, but
before that required a bit of thought.

Fixes #21