tests: read test data directly from the source dir
These used to be copied to the build dir, but they tended to get out of
date, as they were only copied during configuration.

The test data directory is baked into the test executables, so they can
be ran outside of ctest easily.

This adds a function to get the full path of a test data, so if we want
to change this again it should be a bit easier.