HOA: inherit get, has and unset privately in AudioPackFormatHoa
These names refer to both the overloaded tag methods, and the templated
wrapper methods.

The tag methods should be private rather than public (they're defined as
protected in the superclass so that they can be used here), and it's OK
for these using declarations to be private as the wrapper methods are
re-defined.

This was motivated by doxygen, which (incorrectly) shows two definitions
of the templated wrapper methods.