Support 2076-2-style structures in reassignIds (#184)
The approach has been to process ACFs stemming from ATUIDs in the `reassignAudioTrackUidIds`. If an ACF is found, apply an available ACF ID.
The whole "reassignIds" logic has been wrapped in a class, with an `IdIssuer` class used to track IDs and ensure only unique ID's are issued to elements. This avoids the need for lookups and reduces to linear complexity. Note that a common ID value is used for groups of ACF, ATF and ASF elements so that the ADM XML is more readable and easy to follow.