Loosen blockformat ID validation to allow for S-ADM frames
An audioChannelFormat in an S-ADM frame can contain a set of AudioBlockFormats
with an initialiser block.
This is a special block that has a counter value of 0. It may also contain a run
of contiguous blocks that start from some integer n > 1. This commit allows
non-contiguous IDs to be added to a document for this case only.